Add incl_fact_horizon parameter (LuxFloat)
This modification adds a new CLI parameter, 'incl_fact_horizon'
("inclination factor"), which controls the effect the current
inclination has on self-leveling in the Horizon flight mode.  The
current inclination is the number of degrees of pitch or roll that the
vehicle is away from level (whichever is greater).  The modification
only effects PID controller 2 ("LuxFloat").