Added experimental PID controller 6 "Bruce Lee"
PID Controller 6 is an experimental PID controller, which is same as
PID1 for P and I, but with different D gain filtering. D term is the
most sensitive tuning
parameter when it comes to vibrations and PT1 element is implemented,
which acts as a low pass filter to make the Dterm less sensitive to
vibrations.
Therefore this PID controller allows it easily to run less gyro_lpf
filtering from MPU6050 to get less delay to the PID controller and
better tuning for Acro flights.