CLI `rxfail` command now supports ALL channels...
Channels 0-3 (RPYT) can be set to `h` to hold last value, or to `s` to have automatic control. This means RPY to mid stick and T to `rx_min_usec` (or mid stick for 3D mode).
Channels 4..n are the aux channels. These can be set to `h` to hold last value, or to `s` to set a pre-defined value with `rxfail 4 s 1500.

Initialization and MSP messages are adapted to this.