First cut of channel stability.
This is an alternative approach to the 'extra disarm safety' in PR
#1503.

This PR takes the approach of checking the deltas of the last few
channel readings and determining a LOW, MEDIUM or HIGH stability.

This stability value can then be used by all parts of the system, in
particular the code that activates/deactivates mode ranges or
 as well as the code that checks the throttle status
before disarming.