introduce a simple absctraction between the video hardware and the OSD code.
the concept of font mapping, raw characters, cursors, writing strings
and a screen buffer is no-longer in the video driver.

The main OSD code that deals with putting characters in certain
positions and using characters to display raw font symbols is no-longer
video driver dependant.

The font mapping is now located with the font.